Fix issue with base64 termination
- The last frame of base64_encodeData() output was improper
terminated (there was some strange padding with zeros taking place)
but due to base64_decodeData() function seemingly was able to handle
such strangely padded endframes, with base64 command there no
problems appeared.

- But an issue came up now with the crypt64 script (crypt64  includes
and is based on the base64 script and makes usage of base64_encodeData()
and  base64_decodeData()  independently). With this commit the issue
seems fixed and now base64 and as well crypt64 also seem to be fixed
even if this fix is not end solution which still need a bit optimization